+++ annotated_test.go
@@ -430,12 +430,16 @@ func TestAnnotate(t *testing.T) {
 	type a struct{}
 	type b struct{ a *a }
 	type c struct{ b *b }
+	type sliceA struct{ sa []*a }
 	newA := func() *a { return &a{} }
 	newB := func(a *a) *b {
 		return &b{a}
 	}
 	newC := func(b *b) *c {
 		return &c{b}
+	}
+	newSliceA := func(sa ...*a) *sliceA {
+		return &sliceA{sa}
 	}
 
 	t.Run("Provide with optional", func(t *testing.T) {
@@ -492,6 +496,46 @@ func TestAnnotate(t *testing.T) {
 		require.Error(t, err)
 		assert.Contains(t, err.Error(), `missing dependencies`)
 		assert.Contains(t, err.Error(), `missing type: *fx_test.a[name="a"]`)
+	})
+
+	t.Run("Provide with variadic function", func(t *testing.T) {
+		t.Parallel()
+
+		var got *sliceA
+		app := fxtest.New(t,
+			fx.Provide(
+				fx.Annotated{Group: "as", Target: newA},
+				fx.Annotated{Group: "as", Target: newA},
+				fx.Annotate(newSliceA,
+					fx.ParamTags(`group:"as"`),
+				),
+			),
+			fx.Populate(&got),
+		)
+		defer app.RequireStart().RequireStop()
+		require.NoError(t, app.Err())
+
+		assert.Len(t, got.sa, 2)
+	})
+
+	t.Run("Invoke with variadic function", func(t *testing.T) {
+		t.Parallel()
+
+		type T1 struct{ s string }
+
+		app := fxtest.New(t,
+			fx.Supply(
+				fx.Annotate(T1{"foo"}, fx.ResultTags(`group:"t"`)),
+				fx.Annotate(T1{"bar"}, fx.ResultTags(`group:"t"`)),
+				fx.Annotate(T1{"baz"}, fx.ResultTags(`group:"t"`)),
+			),
+			fx.Invoke(fx.Annotate(func(got ...T1) {
+				assert.ElementsMatch(t, []T1{{"foo"}, {"bar"}, {"baz"}}, got)
+			}, fx.ParamTags(`group:"t"`))),
+		)
+
+		defer app.RequireStart().RequireStop()
+		require.NoError(t, app.Err())
 	})
 
 	t.Run("provide with annotated results", func(t *testing.T) {