+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- No changes yet.
+### Changed
+- `fx.Annotate`: support variadic functions, and feeding value groups into
+  them.
 
 ## [1.15.0] - 2021-11-08
 ### Added